refactor: lens-debate — linkNav.ts braids two independent navigation concerns into one module
Split solid-browser linkNav.ts into relativePath.ts (GitHub-relative resolver) and previewPath.ts (codec-injected preview inversion), re-exported via index.ts; split tests to match. No behavior change; 29 tests pass, typecheck clean.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-1, raised by hickey). Not pushed or merged.
