+++ packages/solid-browser/src/index.ts
@@ -1,6 +1,2 @@
-export {
-  type PreviewPathCodec,
-  pathFromPreviewPathname,
-  resolveLinkHref,
-  resolveRelativePath,
-} from "./linkNav";
+export { resolveLinkHref, resolveRelativePath } from "./relativePath";
+export { type PreviewPathCodec, pathFromPreviewPathname } from "./previewPath";
