+++ packages/solid-browser/src/relativePath.ts
@@ -1,14 +1,9 @@
-/** Navigation math for a content browser — pure, framework-free, host-agnostic.
+/** GitHub-style relative-ref resolution — pure, framework-free, host-agnostic.
  *
- *  Two kinds of link a browser must follow:
- *   - a **relative ref** inside rendered prose (a Markdown `[doc](../x.md)` or
- *     `![](logo.png)`) → resolve against the source document, GitHub-style;
- *   - a **preview pathname** reported by a sandboxed iframe after an in-frame
- *     `<a>` click → invert the host's preview-URL encoding back to a path.
- *
- *  Neither knows about git, repos, or kolu — the host's URL codec is injected,
- *  and a "document path" is whatever opaque string the host resolves content
- *  from. */
+ *  Resolves a **relative ref** inside rendered prose (a Markdown `[doc](../x.md)`
+ *  or `![](logo.png)`) against the source document's own directory, the way
+ *  GitHub does. It knows nothing about git, repos, or kolu — a "document path"
+ *  is whatever opaque string the host resolves content from. */
 
 // `/url-policy` is the DOM-free subpath — importing the package root would drag
 // in the `<Markdown>` component (and `solid-js/web`); this stays node-pure.
@@ -42,52 +37,6 @@ export function resolveRelativePath(
 export function resolveLinkHref(fromPath: string, href: string): string | null {
   const path = href.trim().replace(/[?#].*$/, "");
   return resolveRelativePath(fromPath, path);
-}
-
-/** The host's preview-URL codec — how it encodes a document path into the path
- *  segment of a sandboxed-preview URL, and back. Injected because the encoding
- *  is the host's contract (kolu's lives in `kolu-common/preview`), not this
- *  package's. `encode`/`decode` must round-trip. */
-export type PreviewPathCodec = {
-  encode: (path: string) => string;
-  decode: (encoded: string) => string;
-};
-
-/** Map a sandboxed preview's reported `location.pathname` back to the document
- *  path it shows. The preview is served at `<prefix>/<encode(path)>?v=…`; after
- *  an in-frame link click the frame reports its own `location.pathname` (the
- *  opaque-origin sandbox blocks the parent from reading it directly).
- *
- *  The prefix isn't known here — it's derived from the file currently shown:
- *  `currentUrl` ends with `encode(currentPath)`, and everything before that is
- *  the shared prefix. Using the same injected codec for both directions means
- *  the inversion can't drift from the encoding — no second source of truth.
- *
- *  Returns null when the frame navigated outside the preview route (an external
- *  link, or a prefix mismatch) — the caller leaves selection untouched. */
-export function pathFromPreviewPathname(
-  reportedPathname: string,
-  currentUrl: string,
-  currentPath: string,
-  codec: PreviewPathCodec,
-): string | null {
-  const currentPathname = currentUrl.split("?")[0] ?? currentUrl;
-  const encodedCurrent = codec.encode(currentPath);
-  if (!currentPathname.endsWith(encodedCurrent)) return null;
-  const prefix = currentPathname.slice(
-    0,
-    currentPathname.length - encodedCurrent.length,
-  );
-  if (!reportedPathname.startsWith(prefix)) return null;
-  const encodedNext = reportedPathname.slice(prefix.length);
-  if (encodedNext === "") return null;
-  try {
-    return codec.decode(encodedNext);
-  } catch {
-    // A malformed percent-sequence can only arrive if the previewed page
-    // crafted a bogus pathname — treat it as "no navigation".
-    return null;
-  }
 }
 
 /** Directory portion of a path (`"docs/a.md"` → `"docs"`, `"README.md"` → `""`). */
